Created general bayesian optimizing loop that is called by both bayesExp and MOBO
